productos id modified to long

+++ src/main/java/com/markdirect/markdirect/beans/Producto.java
@@ -3,7 +3,7 @@ package com.markdirect.markdirect.beans;
 public class Producto {
 
 	//Atributos
-	int id;
+	long id;
 	String nombre;
 	int id_catNivel2;
 	int id_catNivel1;
@@ -14,7 +14,7 @@ public class Producto {
 	//Constructores
 	public Producto(){}
 	
-	public Producto(int id, String nombre, int id_catNivel2, int id_catNivel1, String marca, int pvp, String imagen) {
+	public Producto(long id, String nombre, int id_catNivel2, int id_catNivel1, String marca, int pvp, String imagen) {
 		super();
 		this.id = id;
 		this.nombre = nombre;
@@ -26,10 +26,10 @@ public class Producto {
 	}
 	
 	//Getters/Setters
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNombre() {
